Concepts
config: custom-environment-variables - used a second layer for protection - if there is same name variable as secret key then that variable is used instead of config key.

schema.methods.customFunction:
we can create custom methods for schema in the model file. it will separate the logics (reusability + code beauty).
we just call it and it will do the required functionality.
This is know as fatty modeling.
But we have to also avoid fatty model.

So learn how to avoid fatty models.

Good luck.